network_anomaly_detection/requirements.txt
# Network Anomaly Detection Requirements

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Configuration Management
PyYAML>=6.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For enhanced GPU monitoring and optimization
# nvidia-ml-py3>=7.352.0  # For GPU memory monitoring

# Development Tools (optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0

# For data processing and scientific computing
scipy>=1.7.0